change to mod_softimer name again, remove mod_threadtimer, fix int overflow.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@2716 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/timers/mod_softtimer/mod_softtimer.c
@@ -45,7 +45,7 @@ static const char modname[] = "mod_softtimer";
 #define MAX_ELEMENTS 1000
 
 struct timer_private {
-    uint32_t reference;
+    uint64_t reference;
 };
 typedef struct timer_private timer_private_t;
